Refactored main to be Migrator.java
