--- precision-medicine/main/test/MainTest.java
@@ -1,12 +1,0 @@
-package grakn.template.java;
-
-import org.junit.Test;
-
-import static org.junit.Assert.assertEquals;
-
-public class MainTest {
-    @Test
-    public void mainShouldRun() {
-        assertEquals("test", "test");
-    }
-}
